// src/db/enums/EuserDetails.ts
export enum EuserDetailsKeys{
    ROLE = "role",
    FIRST_LOGIN = "first_login",
    PICTURE = "picture",
}
export enum EuserDetailsValues{
    /* ruoli */
    SUPERADMIN = "superadmin",
    /*firstLogin */
    NEVER_DID_FIRST_LOGIN = "never_did_first_login",
    HAS_DONE_FIRST_LOGIN = "has_done_first_login"
}